Fix User Dao unit tests
