+++ src/main/java/by/epamtc/digapply/dao/impl/UserDaoImpl.java
@@ -5,7 +5,7 @@ import by.epamtc.digapply.dao.DaoException;
 import by.epamtc.digapply.dao.UserDao;
 import by.epamtc.digapply.entity.User;
 import by.epamtc.digapply.mapper.factory.RowMapperFactory;
-import by.epamtc.digapply.resource.Table;
+import by.epamtc.digapply.dao.Table;
 public class UserDaoImpl extends AbstractDao<User> implements UserDao {
     private static final String SAVE_USER_QUERY = "INSERT INTO Users (user_id, email, password, name, surname, role_id) VALUES (0, ?, ?, ?, ?, ?)";
     private static final String FIND_USER_BY_ID_QUERY = "SELECT * FROM Users WHERE user_id=?";
